# Allow txg_kick() even there is a txg still syncing