--- a/module/zfs/txg.c
+++ b/module/zfs/txg.c
@@ -531,8 +531,6 @@ txg_sync_thread(void *arg)
 		clock_t timeout = zfs_txg_timeout * hz;
 		clock_t timer;
 		uint64_t txg;
-		uint64_t dirty_min_bytes =
-		    zfs_dirty_data_max * zfs_dirty_data_sync_percent / 100;
 
 		/*
 		 * We sync when we're scanning, there's someone waiting
@@ -543,8 +541,7 @@ txg_sync_thread(void *arg)
 		while (!dsl_scan_active(dp->dp_scan) &&
 		    !tx->tx_exiting && timer > 0 &&
 		    tx->tx_synced_txg >= tx->tx_sync_txg_waiting &&
-		    !txg_has_quiesced_to_sync(dp) &&
-		    dp->dp_dirty_total < dirty_min_bytes) {
+		    !txg_has_quiesced_to_sync(dp)) {
 			dprintf("waiting; tx_synced=%llu waiting=%llu dp=%p\n",
 			    tx->tx_synced_txg, tx->tx_sync_txg_waiting, dp);
 			txg_thread_wait(tx, &cpr, &tx->tx_sync_more_cv, timer);
@@ -557,9 +554,11 @@ txg_sync_thread(void *arg)
 		 * prompting it to do so if necessary.
 		 */
 		while (!tx->tx_exiting && !txg_has_quiesced_to_sync(dp)) {
-			if (tx->tx_quiesce_txg_waiting < tx->tx_open_txg+1)
-				tx->tx_quiesce_txg_waiting = tx->tx_open_txg+1;
-			cv_broadcast(&tx->tx_quiesce_more_cv);
+			if (!txg_is_quiescing(dp)) {
+				if (tx->tx_quiesce_txg_waiting < tx->tx_open_txg + 1)
+					tx->tx_quiesce_txg_waiting = tx->tx_open_txg + 1;
+				cv_broadcast(&tx->tx_quiesce_more_cv);
+			}
 			txg_thread_wait(tx, &cpr, &tx->tx_quiesce_done_cv, 0);
 		}
 
@@ -779,6 +778,7 @@ txg_wait_open(dsl_pool_t *dp, uint64_t txg, boolean_t should_quiesce)
  * If there isn't a txg quiescing in the pipeline, push the txg
  * through the pipeline by quiescing the open txg.
  * It is fine there is a txg still syncing.
+ * Pass in the txg number of the transaction that should be closed and synced.
  */
 void
 txg_kick(dsl_pool_t *dp, uint64_t txg)
@@ -787,8 +787,11 @@ txg_kick(dsl_pool_t *dp, uint64_t txg)
 
 	ASSERT(!dsl_pool_config_held(dp));
 
+	if (txg != tx->tx_open_txg ||
+	    tx->tx_quiesce_txg_waiting > tx->tx_open_txg)
+		return;
+
 	mutex_enter(&tx->tx_sync_lock);
-	txg = txg == 0 ? tx->tx_open_txg : txg;
 	if (txg == tx->tx_open_txg &&
 	    !txg_is_quiescing(dp) &&
 	    tx->tx_quiesce_txg_waiting <= tx->tx_open_txg &&