@adamdmoss 2021-04-25
FYI I've been running with this for a couple of days - I haven't benchmarked it but it's been behaving fine so far.