@jxdking 2021-04-27
> In principle this makes good sense to me to help smooth out the txgs.

**Below is my test. Conclusion is at end of this post.**

setup:
KVM client, 12GB Memory, with 8 threads(Haswell)
Ubuntu groovy
8GB zpool on file. The file is on 8GB ramdisk.
Created with
```shell
zpool create -o ashift=12 -O primarycache=metadata test /tmp/mytempfs/d1
```
Parameter
```shell
echo $(expr 1024 \* 1024 \* 256) >> /sys/module/zfs/parameters/zfs_dirty_data_max
echo $(expr 1024 \* 1024 \* 1024) >> /sys/module/zfs/parameters/zfs_arc_min
echo 0 >> /sys/module/zfs/parameters/zfs_abd_scatter_enabled
echo 30 >> /sys/module/zfs/parameters/zfs_dirty_data_sync_percent
```

Note: zfs_dirty_data_sync_percent is set to 30%. I found there is fixed overhead to sync a txg regardless the txg size. Basically, you want to keep txg size as large as possible, so that you can use less number of txgs to get the job done. However the dirty data is soft capped by zfs_delay_min_dirty_percent (60%), and you also want at least 2 txgs in the pipeline at the same time, one for syncing, one for open. Thus, 30% might be the optimal value.

.fio file:
```
; fio-seq-write.job for fiotest

[global]
name=fio-seq-write
filename=/test/fio-seq-write
rw=write
bs=128K
direct=1
numjobs=32
time_based=1
runtime=30
buffer_compress_percentage=70

[file1]
size=4096M
ioengine=libaio
iodepth=16
```

**before patch** 
(I have removed unnecessary lines)
```shell
  WRITE: bw=2312MiB/s (2425MB/s), 62.5MiB/s-93.4MiB/s (65.5MB/s-97.9MB/s), io=67.8GiB (72.8GB), run=30001-30036msec
  WRITE: bw=2364MiB/s (2479MB/s), 63.3MiB/s-100MiB/s (66.3MB/s-105MB/s), io=69.3GiB (74.5GB), run=30001-30037msec
  WRITE: bw=2366MiB/s (2481MB/s), 62.9MiB/s-96.8MiB/s (65.9MB/s-102MB/s), io=69.4GiB (74.5GB), run=30001-30047msec
txg      birth            state ndirty       nread        nwritten     reads    writes   otime        qtime        wtime        stime
34469    44413228261394   C     81412096     0            80900096     0        680      38990172     90459        39504        31888244
34470    44413267251566   C     80756736     0            80183296     0        676      35474837     95056        15563        33482372
34471    44413302726403   C     81543168     0            81043456     0        678      36462139     6058754      20627        26053420
34472    44413339188542   C     81412096     0            80924672     0        679      35180551     5144063      16541        31480321
34473    44413374369093   C     80756736     0            80134144     0        662      41049444     297837       14129        32115820
34474    44413415418537   C     80887808     0            80412672     0        679      35711703     117254       17481        33287954
34475    44413451130240   C     80756736     45056        80269312     1        679      37546015     221847       20722        37446051
34476    44413488676255   C     81412096     0            80912384     0        677      41115482     4964932      44709        28192250
34477    44413529791737   C     81674240     0            81088512     0        681      36332028     114537       23074        31926800
34478    44413566123765   C     81412096     0            80842752     0        680      35750331     145100       39345        34063030
34479    44413601874096   C     81412096     0            80936960     0        686      37437573     124580       13599        34473612
34480    44413639311669   C     81018880     0            80384000     0        669      38736130     122841       34216        34415495
34481    44413678047799   C     81018880     0            80433152     0        673      38211708     219022       10173        32172248
34482    44413716259507   C     81018880     0            80523264     0        673      35520291     232346       18448        31529957
34483    44413751779798   C     80625664     0            80056320     0        672      34776822     194787       13009        32898429
34484    44413786556620   C     81412096     0            80928768     0        677      36494459     108806       18843        32076272
34485    44413823051079   C     80625664     0            80105472     0        670      35260730     59518        32022        31544043
34486    44413858311809   C     80887808     0            80453632     0        675      34858642     5691509      13785        27406610
34487    44413893170451   S     0            0            0            0        0        36041951     206593       22816        0
34488    44413929212402   O     0            0            0            0        0        0            0            0            0
```

**after patch**
```shell
  WRITE: bw=2367MiB/s (2482MB/s), 64.3MiB/s-88.4MiB/s (67.4MB/s-92.7MB/s), io=69.5GiB (74.6GB), run=30001-30068msec
  WRITE: bw=2397MiB/s (2514MB/s), 64.3MiB/s-93.2MiB/s (67.4MB/s-97.7MB/s), io=70.3GiB (75.5GB), run=30001-30029msec
  WRITE: bw=2359MiB/s (2474MB/s), 62.1MiB/s-96.4MiB/s (65.1MB/s-101MB/s), io=69.2GiB (74.3GB), run=30001-30041msec
txg      birth            state ndirty       nread        nwritten     reads    writes   otime        qtime        wtime        stime
38602    45256675295546   C     81281024     0            80781312     0        682      37331366     327246       31073        32381340
38603    45256712626912   C     81281024     0            80658432     0        676      35745088     5685674      42948        26099076
38604    45256748372000   C     81281024     0            80605184     0        672      35411762     7509642      156310       29021598
38605    45256783783762   C     81412096     0            80834560     0        692      39907295     180550       37408        32893259
38606    45256823691057   C     81543168     0            81014784     0        680      36284749     6010436      106064       26588395
38607    45256859975806   C     81412096     0            80736256     0        678      36014189     6982617      27029        23381964
38608    45256895989995   C     81018880     0            80478208     0        678      33998979     244077       43094        30736135
38609    45256929988974   C     81281024     0            80789504     0        682      34759192     5729245      31807        25797541
38610    45256964748166   C     80887808     0            80359424     0        675      34697286     151429       29071        30306521
38611    45256999445452   C     81281024     0            80605184     0        669      33319205     5195582      23199        25388281
38612    45257032764657   C     80887808     0            80236544     0        670      33716458     108571       292652       30580742
38613    45257066481115   C     81281024     0            80629760     0        681      34141174     5460094      213898       25590267
38614    45257100622289   C     81149952     0            80486400     0        678      34771317     8126630      35692        24200981
38615    45257135393606   C     80756736     0            80203776     0        677      35418189     7796445      13185        25106909
38616    45257170811795   C     80887808     0            80248832     0        672      36260117     185389       21869        29588025
38617    45257207071912   C     81149952     0            80596992     0        682      32665857     2732853      66327        29102961
38618    45257239737769   C     81149952     0            80474112     0        669      34803758     3981020      33005        28281408
38619    45257274541527   C     81412096     0            80748544     0        677      35385203     118840       17162        32180451
38620    45257309926730   S     0            0            0            0        0        35474999     114903       26184        0
38621    45257345401729   O     0            0            0            0        0        0            0            0            0
```
Note that the wtime remains small after the patch. There is no txg with "W" status observed, which suggest the fill rate of open txg is slower than sync rate. The open stage is the bottle neck of the whole pipeline. The fio results between "before" and "after" patches are within margin of error. I can say the overhead of extra mutex lock can be ignored. 

 
**I accidentally find "abd scatter" has negative impact on performance by 20 percent. See below.**
after patch (abd scatter enabled)
```shell
  WRITE: bw=1966MiB/s (2062MB/s), 51.9MiB/s-87.4MiB/s (54.4MB/s-91.7MB/s), io=57.7GiB (61.9GB), run=30001-30044msec
  WRITE: bw=1924MiB/s (2017MB/s), 51.1MiB/s-79.6MiB/s (53.5MB/s-83.5MB/s), io=56.5GiB (60.6GB), run=30001-30052msec
  WRITE: bw=1907MiB/s (1999MB/s), 50.6MiB/s-80.2MiB/s (53.1MB/s-84.1MB/s), io=55.9GiB (60.1GB), run=30001-30044msec
txg      birth            state ndirty       nread        nwritten     reads    writes   otime        qtime        wtime        stime
45112    46616059350565   C     81281024     0            80605184     0        675      46312283     219458       14457        42556757
45113    46616105662848   C     81805312     0            81190912     0        684      47335452     359614       21052        41163965
45114    46616152998300   C     81149952     0            80678912     0        679      46887698     12228973     39583        29266544
45115    46616199885998   C     57688064     0            57012224     0        489      12512219     8332132      20771395     42377446
45116    46616212398217   C     81805312     0            81141760     0        683      71568296     3383276      24245        36827690
45117    46616283966513   C     81149952     0            80539648     0        678      44743970     10766099     32691        31341262
45118    46616328710483   C     81149952     0            80527360     0        674      46646751     220928       21371        40313402
45119    46616375357234   C     81149952     0            80527360     0        676      45575267     10715816     20249        29692062
45120    46616420932501   C     81281024     0            80748544     0        680      44971173     236248       25576        42036086
45121    46616465903674   C     81149952     0            80637952     0        680      46693765     214532       25288        43491260
45122    46616512597439   C     81281024     0            80760832     0        676      48310498     96684        6341685      35546532
45123    46616560907937   C     81149952     0            80588800     0        682      46402936     207012       22675        43780165
45124    46616607310873   C     81281024     0            80703488     0        683      48462547     13665952     22716        30395226
45125    46616655773420   C     57294848     0            56705024     0        494      13924883     224172       30015892     39373770
45126    46616669698303   C     101728256    0            101126144    0        835      73702605     4462872      27321        41168582
45127    46616743400908   C     80887808     0            80199680     0        669      55312100     7756103      49028        35153178
45128    46616798713008   C     81281024     0            80719872     0        676      47073669     408426       12489        40645494
45129    46616845786677   C     81149952     0            80637952     0        681      46721243     2543877      106690       38293195
45130    46616892507920   S     0            0            0            0        0        46271377     8687127      33222        0
45131    46616938779297   O     0            0            0            0        0        0            0            0            0
```


Now, I try to slow down the sync stage by adding gzip compression. I put buffer_compress_percentage=70 into .fio file to make sure the compression is kicked in.

**before patch compression=gzip**
```shell
  WRITE: bw=542MiB/s (568MB/s), 13.6MiB/s-23.6MiB/s (14.2MB/s-24.8MB/s), io=16.1GiB (17.2GB), run=30270-30358msec
  WRITE: bw=496MiB/s (520MB/s), 11.0MiB/s-24.9MiB/s (12.5MB/s-26.1MB/s), io=14.6GiB (15.7GB), run=30176-30189msec
  WRITE: bw=507MiB/s (532MB/s), 12.8MiB/s-26.9MiB/s (13.4MB/s-28.2MB/s), io=14.0GiB (16.1GB), run=30182-30191msec
txg      birth            state ndirty       nread        nwritten     reads    writes   otime        qtime        wtime        stime
59754    62524099457587   C     81018880     0            25616384     0        680      563753583    604707       26596        305528904
59755    62524663211170   C     140787712    0            44441600     0        1136     306248761    3026303      22437        497137187
59756    62524969459931   C     81018880     0            25714688     0        700      500277256    156580       1447246      314366264
59757    62525469737187   C     151011328    0            47648768     0        1220     322599153    3290237      30574        526694584
59758    62525792336340   C     80887808     0            25722880     0        679      533332264    4337         57648        299709676
59759    62526325668604   C     141574144    0            44552192     0        1141     299859969    128281       5261         495275433
59760    62526625528573   C     91111424     0            28794880     0        766      498822704    3994640      31961        326383169
59761    62527124351277   C     131088384    0            41373696     0        1075     330498419    146105       11557        464895527
59762    62527454849696   C     91111424     0            29016064     0        783      465150176    150680       9680         344827363
59763    62527919999872   C     130957312    0            41590784     0        1088     345078704    145752       8869         467954698
59764    62528265078576   C     91111424     0            28917760     0        773      468198146    109856       21032        329253132
59765    62528733276722   C     133709824    0            42364928     0        1086     330848546    527090       53951        468862927
59766    62529064125268   C     88489984     0            27975680     0        745      469540196    152291       4839014      321392357
59767    62529533665464   C     133709824    0            42094592     0        1072     326470809    100641       23589        481451266
59768    62529860136273   C     88621056     0            28028928     0        753      481675727    153903       8546         329382950
59769    62530341812000   C     133709824    0            42168320     0        1083     329632969    109922       6349364      471987117
59770    62530671444969   C     92422144     0            29364224     0        758      479269914    95419        22266        338497154
59771    62531150714883   C     137510912    0            43356160     0        1116     340680633    59653        23560        479593707
59772    62531491395516   S     0            0            0            0        0        479781560    106079       24871        0
59773    62531971177076   O     0            0            0            0        0        0            0            0            0
```

**after patch compression=gzip**
```shell
  WRITE: bw=542MiB/s (568MB/s), 14.5MiB/s-22.0MiB/s (15.2MB/s-24.1MB/s), io=15.9GiB (17.1GB), run=30001-30026msec
  WRITE: bw=491MiB/s (515MB/s), 12.9MiB/s-23.8MiB/s (13.5MB/s-24.9MB/s), io=14.4GiB (15.5GB), run=30076-30087msec
  WRITE: bw=506MiB/s (531MB/s), 13.7MiB/s-22.8MiB/s (14.3MB/s-23.0MB/s), io=14.9GiB (15.0GB), run=30088-30103msec
txg      birth            state ndirty       nread        nwritten     reads    writes   otime        qtime        wtime        stime
60351    63292441072166   C     80625664     0            25665536     0        703      297278152    86551        286099131    293857052
60352    63292738350318   C     80625664     0            25432064     0        684      293177147    6684494      280244100    316954118
60353    63293031527465   C     80756736     0            25620480     0        692      298168675    2631230      303171202    289437707
60354    63293329696140   C     80756736     0            25620480     0        690      311718638    93003        283509921    292115530
60355    63293641414778   C     80756736     0            25546752     0        704      292827654    16694        282958337    291062908
60356    63293934242432   C     80756736     0            25473024     0        689      291490036    17600488     265024507    294677914
60357    63294225732468   C     81018880     0            25739264     0        695      290111552    147367       287134821    295117765
60358    63294515844020   C     80756736     0            25473024     0        690      293783917    40360        288662013    292821039
60359    63294809627937   C     80756736     0            25460736     0        679      294224677    37767        287350163    295239715
60360    63295103852614   C     80887808     0            25587712     0        706      292917223    279658       289516676    289092422
60361    63295396769837   C     81018880     0            25579520     0        693      296995015    883183       281102345    306474454
60362    63295693764852   C     80625664     0            25444352     0        668      287265507    65619        301226894    311818433
60363    63295981030359   C     80756736     0            25595904     0        696      319446207    9104382      284642452    303984653
60364    63296300476566   C     81281024     0            25833472     0        691      297675424    59914        300081102    291334517
60365    63296598151990   C     80887808     0            25661440     0        688      304402514    41802        287117406    290045658
60366    63296902554504   C     80756736     0            25620480     0        690      294605822    31144863     251541507    296751848
60367    63297197160326   C     80756736     0            25460736     0        684      287862885    39212        291608968    291407922
60368    63297485023211   S     0            0            0            0        0        302319539    717188       280107850    0
60369    63297787342750   W     0            0            0            0        0        285875039    98920        0            0
60370    63298073217789   O     0            0            0            0        0        0            0            0            0
```
A txg with "W" status is observed. However, there is no difference on performance number.

Then I try lz4. 
**before patch compression=lz4**
```shell
  WRITE: bw=2131MiB/s (2235MB/s), 58.0MiB/s-90.8MiB/s (60.9MB/s-95.2MB/s), io=62.5GiB (67.1GB), run=30001-30009msec
  WRITE: bw=2137MiB/s (2241MB/s), 58.3MiB/s-82.7MiB/s (61.1MB/s-86.7MB/s), io=62.7GiB (67.3GB), run=30001-30033msec
  WRITE: bw=2120MiB/s (2223MB/s), 57.3MiB/s-85.4MiB/s (60.1MB/s-89.6MB/s), io=62.2GiB (66.7GB), run=30001-30017msec
txg      birth            state ndirty       nread        nwritten     reads    writes   otime        qtime        wtime        stime
59497    62295821826121   C     156778496    0            54571008     0        735      53757120     1050416      23781        60254898
59498    62295875583241   C     80887808     0            28323840     0        410      64262346     133175       22957        52541648
59499    62295939845587   C     155992064    0            54206464     0        772      55745896     3385100      28723        63430302
59500    62295995591483   C     80756736     0            28360704     0        573      69832912     293554       40418        55024905
59501    62296065424395   C     154550272    0            53481472     0        718      58163148     4644850      27067        57527882
59502    62296123587543   C     80625664     0            28073984     0        452      64489090     47688        66250        51760139
…